+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/Modules/AzSources.cs
@@ -47,6 +47,31 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         /// <param name="sourceIds">The collection of source IDs to clear.</param>
         ValueTask ClearById(string mapId, IEnumerable<string> sourceIds);
 
+        /// <summary>
+        /// Retrieves a shape from a specific source on the specified map as JS object reference.
+        /// </summary>
+        /// <param name="mapId"></param>
+        /// <param name="sourceId"></param>
+        /// <param name="shapeId"></param>
+        /// <returns></returns>
+        ValueTask<IJSObjectReference?> GetJSShapeById(string mapId, string sourceId, string shapeId);
+
+        /// <summary>
+        /// Retrieves shapes from a specific source on the specified map as JS object references by their IDs.
+        /// </summary>
+        /// <param name="mapId"></param>
+        /// <param name="sourceId"></param>
+        /// <param name="shapeIds"></param>
+        /// <returns></returns>
+        ValueTask<List<IJSObjectReference>> GetJSShapesById(string mapId, string sourceId, IEnumerable<string> shapeIds);
+
+        /// <summary>
+        /// Retrieves the shapes from a specific source on the specified map as JS object references.
+        /// </summary>
+        /// <param name="mapId"></param>
+        /// <param name="sourceId"></param>
+        /// <returns></returns>
+        ValueTask<List<IJSObjectReference>> GetJSShapes(string mapId, string sourceId);
 
         /// <summary>
         /// Retrieves the shapes from a specific source on the specified map.
@@ -140,6 +165,24 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         {
             var module = await moduleTask.Value;
             await module.InvokeVoidAsync(GetJsInteropMethod(), mapId, sourceIds);
+        }
+
+        public async ValueTask<IJSObjectReference?> GetJSShapeById(string mapId, string sourceId, string shapeId)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<IJSObjectReference?>(GetJsInteropMethod(), mapId, sourceId, shapeId);
+        }
+
+        public async ValueTask<List<IJSObjectReference>> GetJSShapesById(string mapId, string sourceId, IEnumerable<string> shapeIds)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<List<IJSObjectReference>?>(GetJsInteropMethod(), mapId, sourceId, shapeIds) ?? [];
+        }
+
+        public async ValueTask<List<IJSObjectReference>> GetJSShapes(string mapId, string sourceId)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<List<IJSObjectReference>?>(GetJsInteropMethod(), mapId, sourceId) ?? [];
         }
 
         public async ValueTask<List<MapEventShape>> GetShapes(string mapId, string sourceId)