+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/Modules/ModuleExtensions.cs
@@ -6,6 +6,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
     {
         Animations,
         Atlas,
+        BoundingBox,
         Common,
         Configuration,
         Controls,
@@ -13,6 +14,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         Factory,
         Features,
         Geolocations,
+        Math,
         Sprites,
         Layers,
         Markers,